requirements.txt
# 🎬 CineAI - AI Video Studio Platform
# Dependências Python essenciais

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8

# Environment Management
python-dotenv==1.0.0

# AI APIs
openai==1.3.6
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0

# Data Processing
Pillow==10.1.0

# Production Server
gunicorn==21.2.0

# Development
python-dateutil==2.8.2